Add client support for basic auth with test and docs
Since we already expose a rack app consumers of this rack
app can use their own implementation of basic auth, which
is basically the out of box support from rack.

What this change does is to ensure that grpc client forwards
that basic auth information to the server when present.

Additionally adds doc and an integration spec