Allow RackApp.handle to work with module and block
This ensures that when a service interface is being passed to handle
it accordingly works/adheres to the same function. Additionally, we
also raise an error when an interface is passed but no block.

This was caught when we were calling the block/proc being passed to handle
(from instrumentation wrapper) and the service classes were getting reloaded
from the initializer. Which is no longer possible start rails 6.

Being able to pass service interface with a block allows us to not reload
classes at the time of init. The additional error raise is future-proofing/guard rails